+++ packages/nano-ui/lib/NanoUI/Monad.hs
@@ -42,6 +42,8 @@ module NanoUI.Monad
   , styled
   , themed
   , disabledWhen
+  , followSystemThemeUi
+  , systemAppearance
 
     -- * Focus and clipboard
   , holdFocus